+++ src/string/is-aws-arn/index.ts
@@ -1,0 +1,2 @@
+export * from './is-aws-arn.decorator'
+export * from './is-aws-arn.predicate'